fix: wear count

+++ PerfumeTracker.xTests/PerfumeSuggestedTests.cs
@@ -20,6 +20,7 @@ public class PerfumeSuggestedTests(WebApplicationFactory<Program> factory) : ICl
 			if (!dbUp) {
 				using var scope = factory.Services.CreateScope();
 				using var context = scope.ServiceProvider.GetRequiredService<PerfumetrackerContext>();
+				if (!context.Database.GetDbConnection().Database.ToLower().Contains("test")) throw new Exception("Live database connected!");
 				var sql = "truncate table \"public\".\"PerfumeWorn\" cascade; truncate table \"public\".\"Perfume\" cascade;  truncate table \"public\".\"PerfumeSuggested\" cascade;";
 				await context.Database.ExecuteSqlRawAsync(sql);
 				context.Perfumes.AddRange(perfumeSeed);
